% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulting.R
\name{create_compositional_barplot}
\alias{create_compositional_barplot}
\title{Compositional Stacked Bar Plot 생성 함수}
\usage{
create_compositional_barplot(
  data,
  time_var,
  comp_vars,
  n_bins = 3,
  title = "Compositional Change by Time"
)
}
\arguments{
\item{data}{분석할 데이터프레임}

\item{time_var}{시간 정보를 담은 연속형 변수 이름 (따옴표 없이)}

\item{comp_vars}{조성 비율 정보를 담은 열 이름들의 벡터 (character vector)}

\item{n_bins}{나눌 시간 구간의 수. 기본값은 3.}

\item{title}{그래프 제목}
}
\description{
Compositional Stacked Bar Plot 생성 함수
}
