% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{perform_and_visualize_clustering}
\alias{perform_and_visualize_clustering}
\title{통합 클러스터링 및 시각화 함수}
\usage{
perform_and_visualize_clustering(data, k, method = "umap")
}
\arguments{
\item{data}{수치형 데이터프레임 또는 행렬.}

\item{k}{원하는 군집의 개수.}

\item{method}{사용할 알고리즘. "kmeans", "hierarchical", "tsne", "umap" 중 선택.}
}
\value{
ggplot 시각화 객체.
}
\description{
지정된 방법과 군집 수(k)로 클러스터링 및 시각화를 수행합니다.
}
