% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packges.R
\name{png.list2call}
\alias{png.list2call}
\title{리스트를 list(name=name, ...) 호출 객체로 변환}
\usage{
png.list2call(input_list, output_type = "call")
}
\arguments{
\item{input_list}{명명된(named) 리스트}

\item{output_type}{반환 타입. "call" (기본값) 또는 "string"}
}
\value{
"call"인 경우 unevaluated call, "string"인 경우 character
}
\description{
리스트를 list(name=name, ...) 호출 객체로 변환
}
