% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{png.plot.scatter.marginal}
\alias{png.plot.scatter.marginal}
\title{Scatter plot with Marginal Density Plots}
\usage{
png.plot.scatter.marginal(
  data,
  x_var,
  y_var,
  group_var,
  palette = "jco",
  size = 3,
  alpha = 0.6,
  rel_widths = c(2, 1),
  rel_heights = c(1, 2)
)
}
\arguments{
\item{data}{A data frame.}

\item{x_var}{String. The name of the variable for the x-axis.}

\item{y_var}{String. The name of the variable for the y-axis.}

\item{group_var}{String. The name of the variable for coloring and grouping.}

\item{palette}{Optional. The color palette (e.g., "jco", "npg", "aaas").}

\item{size}{Optional. Point size for the scatter plot.}

\item{alpha}{Optional. Point alpha transparency for the scatter plot.}
}
\value{
A combined plot object (from cowplot).
}
\description{
Creates a scatter plot with marginal density plots on the top and right.
}
